[feat][sophia22001] Todo 리스트 기능
